Add vmclock backend for VM UBE detection, preferring it over SysGenID
Add support for the non-proprietary vmclock interface (/dev/vmclock0)
for VM UBE detection, preferred over /dev/sysgenid when available and
falling back to it otherwise (V2082724891). SysGenID is retained, not
removed. The generation counter widens to uint64_t and vmclock is read
via the spec's seqlock protocol with bounded retries. Test wiring gains
a per-backend flag so single-backend builds only touch their own
stand-in file; CI builds each backend separately.

+++ crypto/fipsmodule/rand/rand_test.cc
@@ -233,7 +233,7 @@ TEST_F(randTest, UbeDetectionMocked) {
 
   MockedUbeDetection(
     [](uint64_t gn) {
-      set_vm_ube_generation_number_FOR_TESTING(static_cast<uint32_t>(gn));
+      set_vm_ube_generation_number_FOR_TESTING(gn);
     }
   );
 }
